ch01/requirements.txt
Pillow==10.0.1



